requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
click==8.1.7
rich==13.6.0
pytest==7.4.3
colorama==0.4.6